# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API of sowgatly.app",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        },
        {
            "url": "https://sowgatly.app",
            "description": "Remote server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user's orders",
                "operationId": "f641842a2abf975fee5362d88c320539",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search by order number or product name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "delivering",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order from the current cart",
                "operationId": "b0f5daac05dcbd2f54bd761f92afcbf6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "delivery_type",
                                    "recipient_phone"
                                ],
                                "properties": {
                                    "delivery_type": {
                                        "type": "string",
                                        "enum": [
                                            "asap",
                                            "scheduled"
                                        ],
                                        "example": "asap"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "recipient_phone": {
                                        "type": "string",
                                        "example": "65656585"
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Aýna",
                                        "nullable": true
                                    },
                                    "delivery_address": {
                                        "description": "Required unless fulfillment is pickup",
                                        "type": "string",
                                        "example": "Aýtakow köç. 17",
                                        "nullable": true
                                    },
                                    "fulfillment": {
                                        "type": "string",
                                        "default": "delivery",
                                        "enum": [
                                            "delivery",
                                            "pickup"
                                        ]
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "default": "cash",
                                        "enum": [
                                            "cash",
                                            "online"
                                        ]
                                    },
                                    "payment_bank": {
                                        "description": "Required when payment_method is online; a code from GET /api/payment-methods",
                                        "type": "string",
                                        "example": "rysgal",
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel a pending order",
                "operationId": "4bff05b82387c6b20976c7b82ae2c9ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Order is no longer pending"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Legacy alias of GET /api/orders",
                "operationId": "98ff3e4fa7abc93fdf76421a03a40dfd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "operationId": "ca5d23a81020b17dad07d341668d73d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}